Made changes for equipment in frmDataEntry.vb
Made combo boxes in the hire panel suggest from their list of items.
Added a quantity counter to the hire panel.
Saving/Opening/Deleting Equipment Records now is functional (via the ribbon)
The form should now also read in all of the files and their information (equipment only at the moment)
On frmDataEntry(Load), we successfully read in all equipment data and populate the hire combo box.
Added a delete button to the ribbon. I forgot to add it during the initial designs.